PR cleanup third pass
